# [executor] Use sync.Pool to manage free list of ExecuteNodeTask.