This improves performance significantly for large queries on long-running servers. We added Artemis engine to the benchmark built by TravelgateX in https://github.com/travelgateX/presenters-benchmark, the result looks promising,

### Before

```
$ go test -benchmem -cpuprofile artemis65536-before.pprof -test.benchtime 10s -run=^$ . -bench BenchmarkCandidate_artemis_65536_high
goos: darwin
goarch: amd64
pkg: github.com/travelgateX/presenters-benchmark/benchmark
BenchmarkCandidate_artemis_65536_high-8   	      10	1406915638 ns/op	1176622996 B/op	21497498 allocs/op
PASS
ok  	github.com/travelgateX/presenters-benchmark/benchmark	22.516s
```

### After

```
$ go test -benchmem -cpuprofile artemis65536-after.pprof -test.benchtime 10s -run=^$ . -bench BenchmarkCandidate_artemis_65536_high
goos: darwin
goarch: amd64
pkg: github.com/travelgateX/presenters-benchmark/benchmark
BenchmarkCandidate_artemis_65536_high-8   	      10	1270758834 ns/op	878820608 B/op	16844477 allocs/op
PASS
ok  	github.com/travelgateX/presenters-benchmark/benchmark	20.322s
```

### Summary

ns/op: ~9.7% shorter
B/op: ~25.3% less
allocs/op: ~21.6% less